Fix refcount in RawMessageImpl serialization
The headers are built using a pooled buffer, but this was never
released as it was immediately passed into a ByteBufPair and
returned. In DoubleByteBuf this wasn't a problem, as DoubleByteBuf
expected a refcount ownership to be transferred on creation (in fact
there was a problem with headersAndPayload), but it needs to be
released now that we are using ByteBufPair.